# api_specs.json
{"openapi":"3.1.0","info":{"title":"BESESR - Essay Grading Benchmarking Platform","description":"\n    Benchmarking platform for automatic essay grading models. \n    \n    ## New Workflow:\n    1. **Download datasets** using `/datasets/download/all`\n    2. **Run your model** locally on all datasets\n    3. **Submit CSV results** using `/outputs/upload-results`\n    4. **View leaderboard** with your model's performance\n    \n    Submit your model results for evaluation across 12 curated academic datasets.\n    ","version":"1.0.0"},"paths":{"/api/datasets/":{"get":{"tags":["datasets"],"summary":"Get All Datasets","description":"Get all available datasets from dynamic HuggingFace configuration","operationId":"get_all_datasets_api_datasets__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/datasets/{dataset_name}":{"get":{"tags":["datasets"],"summary":"Get Dataset Details","description":"Get detailed information about a specific dataset","operationId":"get_dataset_details_api_datasets__dataset_name__get","parameters":[{"name":"dataset_name","in":"path","required":true,"schema":{"type":"string","title":"Dataset Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/datasets/{dataset_name}/sample":{"get":{"tags":["datasets"],"summary":"Get Dataset Sample","description":"Get a sample of essays from the dataset","operationId":"get_dataset_sample_api_datasets__dataset_name__sample_get","parameters":[{"name":"dataset_name","in":"path","required":true,"schema":{"type":"string","title":"Dataset Name"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/datasets/{dataset_name}/stats":{"get":{"tags":["datasets"],"summary":"Get Dataset Statistics","description":"Get statistics about a dataset","operationId":"get_dataset_statistics_api_datasets__dataset_name__stats_get","parameters":[{"name":"dataset_name","in":"path","required":true,"schema":{"type":"string","title":"Dataset Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/datasets/download/all":{"get":{"tags":["datasets"],"summary":"Download All Datasets","description":"Download all datasets as a ZIP file for local evaluation","operationId":"download_all_datasets_api_datasets_download_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/datasets/download/{dataset_name}":{"get":{"tags":["datasets"],"summary":"Download Single Dataset","description":"Download a single dataset as CSV","operationId":"download_single_dataset_api_datasets_download__dataset_name__get","parameters":[{"name":"dataset_name","in":"path","required":true,"schema":{"type":"string","title":"Dataset Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/datasets/debug/inspect/{dataset_name}":{"get":{"tags":["datasets"],"summary":"Inspect Dataset Structure","description":"Inspect actual dataset structure for column mapping","operationId":"inspect_dataset_structure_api_datasets_debug_inspect__dataset_name__get","parameters":[{"name":"dataset_name","in":"path","required":true,"schema":{"type":"string","title":"Dataset Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/datasets/health/check":{"get":{"tags":["datasets"],"summary":"Dataset Health Check","description":"Check if dataset loading system is working","operationId":"dataset_health_check_api_datasets_health_check_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/datasets/refresh-discovery":{"post":{"tags":["datasets"],"summary":"Refresh Dataset Discovery","description":"Refresh dynamic dataset discovery from HuggingFace","operationId":"refresh_dataset_discovery_api_datasets_refresh_discovery_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/test-load/{dataset_name}":{"get":{"summary":"Test Load Dataset","description":"Test loading sample essays from Hugging Face dataset","operationId":"test_load_dataset_api_test_load__dataset_name__get","parameters":[{"name":"dataset_name","in":"path","required":true,"schema":{"type":"string","title":"Dataset Name"}},{"name":"sample_size","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"Sample Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/leaderboard/":{"get":{"tags":["leaderboard"],"summary":"Get Leaderboard","description":"Get leaderboard rankings from database","operationId":"get_leaderboard_api_leaderboard__get","parameters":[{"name":"metric","in":"query","required":false,"schema":{"type":"string","default":"quadratic_weighted_kappa","title":"Metric"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/leaderboard/datasets/{dataset_name}":{"get":{"tags":["leaderboard"],"summary":"Get Dataset Leaderboard","description":"Get leaderboard for a specific dataset","operationId":"get_dataset_leaderboard_api_leaderboard_datasets__dataset_name__get","parameters":[{"name":"dataset_name","in":"path","required":true,"schema":{"type":"string","title":"Dataset Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/leaderboard/report":{"get":{"tags":["leaderboard"],"summary":"Get Benchmark Report","description":"Get comprehensive benchmark report from database","operationId":"get_benchmark_report_api_leaderboard_report_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/leaderboard/stats":{"get":{"tags":["leaderboard"],"summary":"Get Leaderboard Stats","description":"Get platform statistics from database","operationId":"get_leaderboard_stats_api_leaderboard_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/leaderboard/metrics":{"get":{"tags":["leaderboard"],"summary":"Get Available Metrics","description":"Get list of available evaluation metrics","operationId":"get_available_metrics_api_leaderboard_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/leaderboard/health":{"get":{"tags":["leaderboard"],"summary":"Leaderboard Health","description":"Health check for leaderboard service","operationId":"leaderboard_health_api_leaderboard_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/outputs/submissions/upload-results":{"post":{"tags":["submissions"],"summary":"Upload Results","description":"Upload CSV results from model evaluation on downloaded datasets","operationId":"upload_results_api_outputs_submissions_upload_results_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_results_api_outputs_submissions_upload_results_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/outputs/submissions/template":{"get":{"tags":["submissions"],"summary":"Get Submission Template","description":"Get template showing expected CSV format","operationId":"get_submission_template_api_outputs_submissions_template_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint with database and dataset status","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/test-openrouter":{"get":{"summary":"Test Openrouter","description":"Test OpenRouter connection (legacy)","operationId":"test_openrouter_test_openrouter_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/debug/database":{"get":{"summary":"Debug Database","description":"Debug endpoint to check database status","operationId":"debug_database_debug_database_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/debug/datasets":{"get":{"summary":"Debug Datasets","description":"Debug endpoint to check dataset loading","operationId":"debug_datasets_debug_datasets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/debug/reset-database":{"post":{"summary":"Reset Database","description":"Reset database (development only)","operationId":"reset_database_debug_reset_database_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api-info":{"get":{"summary":"Api Info","description":"Information about the new API workflow","operationId":"api_info_api_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_results_api_outputs_submissions_upload_results_post":{"properties":{"model_name":{"type":"string","title":"Model Name"},"submitter_name":{"type":"string","title":"Submitter Name"},"submitter_email":{"type":"string","title":"Submitter Email"},"model_description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model Description"},"paper_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Paper Url"},"results_files":{"items":{"type":"string","format":"binary"},"type":"array","title":"Results Files"}},"type":"object","required":["model_name","submitter_name","submitter_email","results_files"],"title":"Body_upload_results_api_outputs_submissions_upload_results_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}